+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - Unreleased
+
+### Added
+
+- `Deserialize` methods for `TomlSerializer` static class
+- Unit tests for `Deserialize` methods
+- Documentation for deserialization rules
+- Documentation for `Deserialize` methods
+
+### Fixed
+
+- Escape double quotes and backslashes in serialized strings
+
 ## [0.2.0] - 2022-12-21
 
 ### Added